initial support for atlassian
